+++ packages/harpguru-core/src/components/covariant-menu/covariant-menu.tsx
@@ -89,7 +89,7 @@ export const CovariantMenu = (menuProps: MenuProps): React.ReactElement => {
         </OptionLock>
       </MenuFace>
       <MenuOpenButton {...menuProps}>
-        <Feather name="sliders" size={sizes['7']} color={colors.pageColor} />
+        <Feather name="sliders" size={sizes['7']} color={colors.inertOutline} />
       </MenuOpenButton>
     </Menu>
   )