+++ packages/harpguru-core/src/types/index.ts
@@ -50,6 +50,7 @@ export type MenuProps = {
   readonly isMenuStashed: boolean
   readonly isLabelHidden: boolean
   readonly stashPosition: MenuStashPosition
+  readonly overrideBackgroundColor?: boolean
   readonly openCloseMenu: () => void
 }